Rename import-session-sheet.pure.ts to import-session-sheet-view-model.ts
